Bugfix: stack_init may corrupt the stack
On startup, the stack is filled with a canary value. This can later be
used to check how much of the stack is in use (see stacktools.h).

However, the startup code itself also needs some stack memory to
operate. stack_init() would also overwrite the stack memory that was
already in use, resulting in undefined behaviour.